components/script/dom/bindings/codegen/test/test_traceProtos.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=744772
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 744772</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=744772">Mozilla Bug 744772</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 744772 **/

SimpleTest.waitForExplicitFinish();

function callback() {
  new XMLHttpRequest().upload;
  ok(true, "Accessing unreferenced DOM interface objects shouldn't crash");
  SimpleTest.finish();
}

delete window.XMLHttpRequestUpload;
SpecialPowers.exactGC(window, callback);

</script>
</pre>
</body>
</html>
